+++ src/index.tsx
@@ -1,12 +1,33 @@
 import React, { useRef, useEffect } from 'react'
-import PropTypes from 'prop-types'
-import { useSprings, animated } from 'react-spring'
+import { useSprings, animated, SpringConfig } from 'react-spring'
 import { useDrag } from 'react-use-gesture'
 import useResizeObserver from 'use-resize-observer'
-import 'intersection-observer'
+
+type SliderProps = {
+  children: React.ReactNode[]
+  index: number
+  onIndexChange: (newIndex: number) => void
+  className?: string
+  style?: {}
+  slideStyle?: React.CSSProperties | ((index: number) => React.CSSProperties)
+  onDragStart?: (pressedIndex: number) => void
+  onDragEnd?: (pressedIndex: number) => void
+  onTap?: (pressedIndex: number) => void
+} & typeof defaultProps
+
+const defaultProps = {
+  enabled: true,
+  vertical: false,
+  slideAlign: 'center',
+  draggedScale: 1,
+  draggedSpring: { tension: 1200, friction: 40 } as SpringConfig,
+  trailingSpring: { tension: 120, friction: 30 } as SpringConfig,
+  releaseSpring: { tension: 120, friction: 30 } as SpringConfig,
+  trailingDelay: 50
+}
 
 // style for the slides wrapper
-const slidesWrapperStyle = vertical => ({
+const slidesWrapperStyle = (vertical: boolean): React.CSSProperties => ({
   display: 'flex',
   height: '100%',
   width: '100%',
@@ -20,9 +41,10 @@ const slidesWrapperStyle = vertical => ({
   touchAction: vertical ? 'pan-x' : 'pan-y'
 })
 
-const clamp = (num, clamp, higher) => Math.min(Math.max(num, clamp), higher)
+const clamp = (num: number, clamp: number, higher: number) =>
+  Math.min(Math.max(num, clamp), higher)
 
-export function Slider({
+export const Slider = ({
   children,
   index,
   onIndexChange,
@@ -30,20 +52,23 @@ export function Slider({
   style,
   slideStyle,
   enabled,
-  trail,
   vertical,
+  slideAlign,
   draggedScale,
   draggedSpring,
+  releaseSpring,
   trailingSpring,
   trailingDelay,
   onDragStart,
-  onDragEnd
-}) {
-  const slideStyleFunc = typeof slideStyle === 'function' ? slideStyle : () => slideStyle
+  onDragEnd,
+  onTap
+}: SliderProps) => {
+  const slideStyleFunc =
+    typeof slideStyle === 'function' ? slideStyle : () => slideStyle
   // root holds are slides wrapper node and we use a ResizeObserver
   // to observe its size in order to recompute the slides position
   // when it changes
-  const root = useRef()
+  const root = useRef<HTMLInputElement>(null)
   const { width, height } = useResizeObserver({ ref: root })
 
   const axis = vertical ? 'y' : 'x'
@@ -60,7 +85,7 @@ export function Slider({
   // visibleIndexes is a Set holding the index of slides that are
   // currently partially or fully visible (intersecting) in the
   // viewport
-  const visibleIndexes = useRef(new Set())
+  const visibleIndexes = useRef(new Set<number>())
   const firstVisibleIndex = useRef(0)
   const lastVisibleIndex = useRef(0)
 
@@ -71,44 +96,44 @@ export function Slider({
 
   // callback called by the intersection observer updating
   // visibleIndexes
-  const cb = slides => {
+  const cb: IntersectionObserverCallback = slides => {
     slides.forEach(({ isIntersecting, target }) =>
-      visibleIndexes.current[isIntersecting ? 'add' : 'delete'](~~target.attributes['data-index'].value)
+      visibleIndexes.current[isIntersecting ? 'add' : 'delete'](
+        Number(target.getAttribute('data-index'))
+      )
     )
     const visibles = Array.from(visibleIndexes.current).sort()
     firstVisibleIndex.current = visibles[0]
     lastVisibleIndex.current = visibles[visibles.length - 1]
   }
 
-  const observer = useRef(null)
+  const observer = useRef<IntersectionObserver | null>(null)
 
   // we add the slides to the IntersectionObserver:
   // this is recomputed everytime the user adds or removes a slide
   useEffect(() => {
     if (!observer.current) observer.current = new IntersectionObserver(cb)
-    Array.from(root.current.children).forEach(t => observer.current.observe(t))
+    Array.from(root.current!.children).forEach(t =>
+      observer.current!.observe(t)
+    )
   }, [children.length, root])
 
   // removing the observer on unmount
-  useEffect(() => () => observer.current.disconnect(), [])
+  useEffect(() => () => observer.current!.disconnect(), [])
 
   // setting the springs with initial position set to restPos:
   // this is important when adding slides since changing children
   // length recomputes useSprings
-  const [springs, set] = useSprings(
-    children.length,
-    i => {
-      // zIndex will make sure the dragged slide stays on top of the others
-      return {
-        x: vertical ? 0 : restPos.current,
-        y: vertical ? restPos.current : 0,
-        s: 1,
-        zIndex: 0,
-        immediate: key => key === 'zIndex'
-      }
-    },
-    [children.length]
-  )
+  const [springs, set] = useSprings(children.length, _i => {
+    // zIndex will make sure the dragged slide stays on top of the others
+    return {
+      x: vertical ? 0 : restPos.current,
+      y: vertical ? restPos.current : 0,
+      s: 1,
+      zIndex: 0,
+      immediate: key => key === 'zIndex'
+    }
+  })
 
   // everytime the index changes, we should calculate the right position
   // of the slide so that its centered: this is recomputed everytime
@@ -120,10 +145,14 @@ export function Slider({
     // here we take the selected slide
     // and calculate its position so its centered in the slides wrapper
     if (vertical) {
-      const { offsetTop, offsetHeight } = root.current.children[index]
+      const { offsetTop, offsetHeight } = root.current!.children[
+        index
+      ] as HTMLElement
       restPos.current = Math.round(-offsetTop + (height - offsetHeight) / 2)
     } else {
-      const { offsetLeft, offsetWidth } = root.current.children[index]
+      const { offsetLeft, offsetWidth } = root.current!.children[
+        index
+      ] as HTMLElement
       restPos.current = Math.round(-offsetLeft + (width - offsetWidth) / 2)
     }
     // two options then:
@@ -131,7 +160,15 @@ export function Slider({
     // is equal to index, we just want to set the position where it should
 
     if (indexRef.current === index) {
-      set(() => ({ [axis]: restPos.current, s: 1, config: key => key === axis && { ...trailingSpring, velocity: velocity.current } }))
+      set(_i => ({
+        [axis]: restPos.current,
+        s: 1,
+        config: { ...releaseSpring, velocity: velocity.current }
+        // config: key =>
+        //   key === axis
+        //     ? { ...releaseSpring, velocity: velocity.current }
+        //     : undefined,
+      }))
     } else {
       // 2. the user has changed the index props: in that case indexRef
       // is outdated and different from index. We want to animate depending
@@ -141,34 +178,58 @@ export function Slider({
       const dir = index < indexRef.current ? -1 : 1
       // if direction is 1 then the first slide to animate should be the lowest
       // indexed visible slide, if -1 the highest
-      const firstToMove = dir > 0 ? firstVisibleIndex.current : lastVisibleIndex.current
+      const firstToMove =
+        dir > 0 ? firstVisibleIndex.current : lastVisibleIndex.current
       set(i => {
         return {
           [axis]: restPos.current,
           s: 1,
-          config: key => key === axis && trailingSpring,
-          delay: i * dir < firstToMove * dir ? 0 : Math.abs(firstToMove - i) * trailingDelay
+          // config: key => key === axis && releaseSpring,
+          config: releaseSpring,
+          delay:
+            i * dir < firstToMove * dir
+              ? 0
+              : Math.abs(firstToMove - i) * trailingDelay
         }
       })
     }
     // finally we update indexRef to match index
     indexRef.current = index
-  }, [index, set, root, vertical, axis, height, width, trail, trailingSpring, draggedSpring, trailingDelay])
+  }, [
+    index,
+    set,
+    root,
+    vertical,
+    axis,
+    height,
+    width,
+    releaseSpring,
+    draggedSpring,
+    trailingDelay
+  ])
 
   // adding the bind listener
   const bind = useDrag(
     ({
       first,
       last,
+      tap,
       vxvy: [vx, vy],
       delta: [dx, dy],
+      swipe: [sx, sy],
       movement: [movX, movY],
       args: [pressedIndex],
       memo = springs[pressedIndex][axis].getValue()
     }) => {
+      if (tap) {
+        onTap && onTap(pressedIndex)
+        return
+      }
       const v = vertical ? vy : vx
       const dir = -Math.sign(vertical ? dy : dx)
       const mov = vertical ? movY : movX
+      const swipe = vertical ? sy : sx
+
       if (first) {
         // if this is the first drag event, we're trailing the controllers
         // to the index being dragged and setting zIndex, scale and config
@@ -176,7 +237,10 @@ export function Slider({
           return {
             [axis]: memo + mov,
             s: draggedScale,
-            config: key => (key === axis && i === pressedIndex ? draggedSpring : trailingSpring),
+            config: key =>
+              key === axis && i === pressedIndex
+                ? draggedSpring
+                : trailingSpring,
             zIndex: i === pressedIndex ? 10 : 0
           }
         })
@@ -186,8 +250,12 @@ export function Slider({
       } else if (last) {
         // when the user releases the drag and the distance or speed are superior to a threshold
         // we update the indexRef
-        if (Math.abs(mov) > size / 2 || Math.abs(v) > 0.3) {
-          indexRef.current = clamp(indexRef.current + (mov > 0 ? -1 : 1), 0, children.length - 1)
+        if (Math.abs(mov) > size! / 2 || swipe !== 0) {
+          indexRef.current = clamp(
+            indexRef.current - swipe,
+            0,
+            children.length - 1
+          )
         }
         // if the index is not equal to indexRef we know we've moved a slide
         // so we tell the user to update its index in the next tick and our useEffect
@@ -199,7 +267,13 @@ export function Slider({
           requestAnimationFrame(() => onIndexChange(indexRef.current))
         }
         // if the index hasn't changed then we set the position back to where it should be
-        else set(() => ({ [axis]: restPos.current, s: 1, config: key => key === axis && trailingSpring }))
+        else
+          set(() => ({
+            [axis]: restPos.current,
+            s: 1,
+            // config: key => key === axis && releaseSpring,
+            config: releaseSpring
+          }))
 
         // triggering onDragEnd prop if it exists
         onDragEnd && onDragEnd(pressedIndex)
@@ -207,12 +281,19 @@ export function Slider({
 
       // if not we're just dragging and we're just updating the position
       else {
-        const firstToMove = dir > 0 ? firstVisibleIndex.current : lastVisibleIndex.current
+        const firstToMove =
+          dir > 0 ? firstVisibleIndex.current : lastVisibleIndex.current
         set(i => {
           return {
             [axis]: mov + memo,
-            delay: i * dir < firstToMove * dir || i === pressedIndex ? 0 : Math.abs(firstToMove - i) * trailingDelay,
-            config: key => (key === axis && i === pressedIndex ? draggedSpring : trailingSpring)
+            delay:
+              i * dir < firstToMove * dir || i === pressedIndex
+                ? 0
+                : Math.abs(firstToMove - i) * trailingDelay,
+            config: key =>
+              key === axis && i === pressedIndex
+                ? draggedSpring
+                : trailingSpring
           }
         })
       }
@@ -220,7 +301,7 @@ export function Slider({
       // and returning memo to keep the initial position in cache along drag
       return memo
     },
-    { enabled, axis }
+    { enabled, axis, filterTaps: true }
   )
 
   return (
@@ -234,12 +315,12 @@ export function Slider({
             key={i}
             data-index={i}
             style={{
+              alignItems: slideAlign,
               ...slideStyleFunc(i),
               zIndex,
               [axis]: pos,
               scale: s,
               display: 'flex',
-              alignItems: 'center',
               willChange: 'transform'
             }}
           >
@@ -251,33 +332,4 @@ export function Slider({
   )
 }
 
-Slider.defaultProps = {
-  className: undefined,
-  style: undefined,
-  enabled: true,
-  trail: true,
-  vertical: false,
-  draggedScale: 1,
-  draggedSpring: { tension: 1200, friction: 40 },
-  trailingSpring: { mass: 1, tension: 280, friction: 60 },
-  trailingDelay: 50,
-  onDragStart: undefined,
-  onDragEnd: undefined
-}
-
-Slider.propTypes = {
-  children: PropTypes.node.isRequired,
-  index: PropTypes.number.isRequired,
-  onIndexChange: PropTypes.func.isRequired,
-  className: PropTypes.string,
-  style: PropTypes.object,
-  slideStyle: PropTypes.object,
-  enabled: PropTypes.bool,
-  vertical: PropTypes.bool,
-  trail: PropTypes.bool,
-  draggedScale: PropTypes.number,
-  draggedSpring: PropTypes.object,
-  trailingSpring: PropTypes.object,
-  onDragStart: PropTypes.func,
-  onDragEnd: PropTypes.func
-}
+Slider.defaultProps = defaultProps